requirements.txt
# AI Music Recommendation System - Cloud Run Deployment
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud services
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.0
google-cloud-run==0.10.0
google-api-python-client==2.108.0

# Elasticsearch
elasticsearch==8.11.0

# Machine learning & data processing
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.5.2
pandas==2.2.3

# HTTP and networking
requests==2.31.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0